requirements.txt
#pandas
#google
langchain
langchain_google_genai
#langchain_community
#langchain_core
#langchain-experimental
#langgraph
#langsmith
#langchainhub
#wikipedia
#umap
#hdbscan
#sentence_transformers
#scikit-learn
#bertopic
#matplotlib
#kaleido
#datasets
#evaluate
python-dotenv
accelerate>=0.26.0
pypdf
faiss-cpu
streamlit